Add council websites, split update into separate scripts
