+++ requirements.txt
@@ -1,3 +1,4 @@
+beautifulsoup4
 pip
 csvs-to-sqlite
 datasette
